refactored login hook to HOC
