+++ src/App.tsx
@@ -1,10 +1,11 @@
-import React, { FC } from 'react';
+import React, { FC, useEffect } from 'react';
 import { ThemeProvider } from 'styled-components';
 import { ConnectedRouter } from 'connected-react-router';
 import { Redirect, Route, Switch } from 'react-router-dom';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 import { selectUserPending } from 'store/userProfile/selectors';
+import ActionTypes from 'store/auth/actionTypes';
 
 import { SignUpWithData } from 'pages/SignUp';
 import { Leaderboard } from 'pages/LeaderBoard';
@@ -14,23 +15,32 @@ import { Forum } from 'pages/Forum';
 import { HomePage } from 'pages/HomePage';
 import { SignInWithData } from 'pages/SignIn';
 import { UserPageWithRouter } from 'pages/UserPage';
-import useIsLoggedIn from 'pages/SignIn/useIsLoggedIn';
 import { ChangePasswordPageWithRouter } from 'pages/ChangePasswordPage';
 import { ChangeUserInfoPageWithRouter } from 'pages/ChangeUserInfoPage';
 
 import ErrorBoundary from 'components/ErrorBoundary/ErrorBoundary';
 
+import ProtectedRoute from 'utils/hocs/protectedRoute';
 import history from 'utils/history';
 
 import { GlobalStyles } from 'ui/global';
 import { themes } from 'ui/themes';
 
 const App: FC = () => {
-    useIsLoggedIn();
+    const dispatch = useDispatch();
+    useEffect(() => {
+        dispatch({ type: ActionTypes.GetUser });
+    }, [dispatch]);
 
     const getUserInfoPending = useSelector(selectUserPending);
 
     const forPending = getUserInfoPending ? 'pending' : '';
+
+    const ProtectedUser = ProtectedRoute(UserPageWithRouter);
+    const ProtectedChangeInfo = ProtectedRoute(ChangeUserInfoPageWithRouter);
+    const ProtectedChangepassword = ProtectedRoute(
+        ChangePasswordPageWithRouter,
+    );
 
     return (
         <ThemeProvider theme={themes.light}>
@@ -64,13 +74,13 @@ const App: FC = () => {
                                     <Forum />
                                 </Route>
                                 <Route exact path="/user">
-                                    <UserPageWithRouter />
+                                    <ProtectedUser />
                                 </Route>
                                 <Route path="/user/change-password">
-                                    <ChangePasswordPageWithRouter />
+                                    <ProtectedChangepassword />
                                 </Route>
                                 <Route path="/user/change-info">
-                                    <ChangeUserInfoPageWithRouter />
+                                    <ProtectedChangeInfo />
                                 </Route>
                                 <Redirect to="/" />
                             </Switch>
